details

+++ app/Http/Controllers/Admin/PlanController.php
@@ -60,14 +60,23 @@ class PlanController extends Controller
     public function destroy($url)
     {
 
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+                        ->with('details')
+                        ->where('url', $url)->first();
 
         if(!$plan)
             return redirect()->back();
-        
+
+        if($plan->details->count() > 0){
+            return redirect()
+            ->back()
+            ->with('error', 'Existem  detalhes vinculados a esse plano, portanto não pode deletar');
+        }
+
         $plan->delete();
 
-        return redirect()->route('plans.index');
+        return redirect()->route('plans.index')
+                        ->with('message', 'Registro deletado com sucesso');
     }
 
     public function search(Request $request)
